==> Task 3.2/Med&BS_to_FDI/04_C_DISCARDS_AT_AGE.Rmd <==
---
title: 'Annex I.4 - STREAM: Auxiliary scripts for the conversion from DG MARE Med&BS
  format to DG MARE FDI format'
author: "M.T. Facchini and I. Bitetto"
date: '`r date()`'
output: word_document
---


```{r compil,eval=F,include=F,cache=F,echo=F,warning=F,message=F,progress=F,verbose=F}
#just a reminder to how compile a docx doc
if(F){
rmarkdown::render("Data.Rmd",c("word_document"))
}
```


```{r results='hide', message=FALSE, warning=FALSE, echo=F}
#R general option:
options(stringsAsFactors = FALSE)

#chunk option
knitr::opts_chunk$set(cache=TRUE,echo=TRUE, warning=FALSE, 
	message=FALSE, fig.height=6,progress=FALSE,verbose=FALSE,
		include=TRUE,dev='png',autodep=FALSE)

#load packages
library(reshape2)
library(reshape)
library(dplyr)
library(knitr)
library(pander)

#pander options
panderOptions('table.split.table', 60)
panderOptions('table.style', 'grid')
panderOptions('table.split.cells', 10)
panderOptions('table.alignment.default', 'left')
panderOptions('table.alignment.rownames', 'right')
panderOptions('decimal.mark', ',')
panderOptions('graph.fontsize', '10')

```


# script 04: *C_DISCARDS_AT_AGE*

This script allows to create the DG MARE FDI C_DISCARDS_AT_AGE starting from DG MARE Med&BS A_Catch table and the communication table used for the conversion of the FISHERY codes.

##Settings 

```{r data1}

# set the working directory
myWD <- paste("C:\\Users\\Bitetto Isabella\\OneDrive - Coispa Tecnologia & Ricerca S.C.A.R.L\\MARE22\\STREAM\\FINAL REVISION OF DELIVERABLES\\DG_MARE_MedBS_to_FDI", sep="")
setwd(myWD)

lev5 <- read.csv("./communicationTable_lev5.csv", sep=";")
loca_A <- read.csv("./A_Catch_example.csv", sep=";")  
JRC_FDI_Table_E <- read.csv("./E_LANDINGS_AT_AGE.csv", sep=";")
template_C <- read.csv("./TABLE_C_MBS_DISCARDS_AGE.csv", sep=";")
  
loca_YEAR <- 9999
loca_GSA <- "99"

```

##Input Data


1) *DG MARE Med&BS A_Catch*

```{r data_in4, echo=F}
pander(head(loca_A[,c(1:24)]))
```


2) *Communication table for the FISHERY codes*

```{r data_in5, echo=F}
pander(lev5)
```


3) *DG MARE FDI C_DISCARDS_AT_AGE template*

```{r data_in6, echo=F}
pander(data.frame(C_DISCARDS_AT_AGE=colnames(template_C)))
```


##Processing tables 

Using some data in the DG MARE Med&BS format :


```{r data2}

# Extraction of numbers, weights and lengths for DISCARDS
loca_A_DIS <-  loca_A[,c(1:24)]

# NUMBERS
Discard_nb_age =  loca_A_DIS
for (i in 1:20){
  Discard_nb_age = cbind(Discard_nb_age,loca_A[,colnames(loca_A)==paste("AGE_",i-1,
                                                                        "_NO_DISCARD",sep="")])
}
Discard_nb_age = cbind(Discard_nb_age,loca_A[,colnames(loca_A)==
                                               "AGE_20_PLUS_NO_DISCARD"])
colnames(Discard_nb_age)=c(colnames(loca_A[1:24]),
                           paste("AGE_",c(0:19),"_NO_DISCARD",sep=""),"AGE_20_PLUS_NO_DISCARD")

Discard_nb_age <- Discard_nb_age[, !( colnames(Discard_nb_age) %in% 
                                        c("NO_LENGTH_MEASUREMENTS_DISCARDS",  "NO_SAMPLES_CATCH", 
                                          "NO_LENGTH_MEASUREMENTS_CATCH", "NO_AGE_MEASUREMENTS_CATCH", 
                                          "NO_LENGTH_MEASUREMENTS_LANDINGS", "SPECON", "ID",
                                          "NO_SAMPLES_LANDINGS", "NO_AGE_MEASUREMENTS_LANDINGS",
                                          "LANDINGS") )]  

DIS_NO <- melt(Discard_nb_age, id=c( "COUNTRY", "YEAR" , 
                                     "QUARTER" , "VESSEL_LENGTH" ,"GEAR"  ,"MESH_SIZE_RANGE",
                                     "FISHERY" , "AREA"  , "SPECIES" , "DISCARDS",
                                     "NO_SAMPLES_DISCARDS", "NO_AGE_MEASUREMENTS_DISCARDS" , 
                                     "MIN_AGE", "MAX_AGE" ))
colnames(DIS_NO)[ncol(DIS_NO)] <- "number"

DIS_NO$variable <- as.character(DIS_NO$variable)
DIS_NO$variable <-   apply(DIS_NO, 1, function(x) 
  ifelse(nchar(x[15]) == 16, substring(x[15], 5, 5) , substring(x[15], 5, 6)))

DIS_NO$variable <- as.numeric(as.character(DIS_NO$variable))
DIS_NO$number <- as.numeric(as.character(DIS_NO$number))


# INDIVIDUAL WEIGHTS
Discard_wei_age =  loca_A_DIS
for (i in 1:20){                                                       
  Discard_wei_age = cbind(Discard_wei_age,loca_A[,colnames(loca_A)==
                                                   paste("AGE_",i-1,"_MEAN_WEIGHT_DISCARD",sep="")])
}
Discard_wei_age = cbind(Discard_wei_age,loca_A[,colnames(loca_A)==
                                                 "AGE_20_PLUS_MEAN_WEIGHT_DISCARD"]) 

colnames(Discard_wei_age)=c(colnames(loca_A[1:24]),
                            paste("AGE_",c(0:19),"_MEAN_WEIGHT_DISCARD",sep=""),
                            "AGE_20_PLUS_MEAN_WEIGHT_DISCARD")

Discard_wei_age <- Discard_wei_age[, !( colnames(Discard_wei_age) %in% 
                                          c( "NO_SAMPLES_DISCARDS", "NO_SAMPLES_LANDINGS", "NO_SAMPLES_CATCH",
                                             "NO_LENGTH_MEASUREMENTS_DISCARDS", "NO_LENGTH_MEASUREMENTS_LANDINGS",
                                             "NO_LENGTH_MEASUREMENTS_CATCH",
                                             "NO_AGE_MEASUREMENTS_DISCARDS","NO_AGE_MEASUREMENTS_LANDINGS", 
                                             "NO_AGE_MEASUREMENTS_CATCH",
                                             "SPECON", "ID", "LANDINGS","DISCARDS","MIN_AGE", "MAX_AGE" ) )]  

DIS_WEI <- melt(Discard_wei_age, id=c( "COUNTRY", "YEAR" , "QUARTER" , 
                                       "VESSEL_LENGTH" ,"GEAR"  ,"MESH_SIZE_RANGE"  ,"FISHERY" , "AREA"  , 
                                       "SPECIES" ))
colnames(DIS_WEI)[ncol(DIS_WEI)] <- "ind_weight"

DIS_WEI$variable <- as.character(DIS_WEI$variable)

DIS_WEI$variable <-   apply(DIS_WEI, 1, function(x) 
  ifelse(nchar(x[10]) == 25, substring(x[10], 5, 5) , substring(x[10], 5, 6)))

DIS_WEI$variable <- as.numeric(as.character(DIS_WEI$variable))
DIS_WEI$ind_weight <- as.numeric(as.character(DIS_WEI$ind_weight))

DIS_NO_WEI <- merge(DIS_NO, DIS_WEI, 
                    by=colnames(DIS_WEI)[colnames(DIS_WEI) != "ind_weight"], all=T )
DIS_NO_WEI$ind_weight[DIS_NO_WEI$ind_weight == -1] <- 0

# INDIVIDUAL LENGTHS
Discard_len_age =  loca_A_DIS
for (i in 1:20){                                                      
  Discard_len_age = cbind(Discard_len_age,loca_A[,colnames(loca_A)==
                                                   paste("AGE_",i-1,"_MEAN_LENGTH_DISCARD",sep="")])
}
Discard_len_age = cbind(Discard_len_age,loca_A[,colnames(loca_A)==
                                                 "AGE_20_PLUS_MEAN_LENGTH_DISCARD"])

colnames(Discard_len_age)=c(colnames(loca_A[1:24]),
                            paste("AGE_",c(0:19),"_MEAN_LENGTH_DISCARD",sep=""),"AGE_20_PLUS_MEAN_LENGTH_DISCARD")

Discard_len_age <- Discard_len_age[, !( colnames(Discard_len_age) %in% 
                                          c("NO_SAMPLES_DISCARDS", "NO_SAMPLES_LANDINGS", "NO_SAMPLES_CATCH",
                                            "NO_LENGTH_MEASUREMENTS_DISCARDS", "NO_LENGTH_MEASUREMENTS_LANDINGS",
                                            "NO_LENGTH_MEASUREMENTS_CATCH",
                                            "NO_AGE_MEASUREMENTS_DISCARDS","NO_AGE_MEASUREMENTS_LANDINGS", 
                                            "NO_AGE_MEASUREMENTS_CATCH",
                                            "SPECON", "ID", "LANDINGS","DISCARDS","MIN_AGE", "MAX_AGE" ) )]  

DIS_LEN <- melt(Discard_len_age, id=c( "COUNTRY", "YEAR" , "QUARTER" , 
                                       "VESSEL_LENGTH" ,"GEAR"  ,"MESH_SIZE_RANGE"  ,"FISHERY" , "AREA"  , 
                                       "SPECIES" ))
colnames(DIS_LEN)[ncol(DIS_LEN)] <- "ind_length"

DIS_LEN$variable <- as.character(DIS_LEN$variable)

DIS_LEN$variable <-   apply(DIS_LEN, 1, function(x) 
  ifelse(nchar(x[10]) == 25, substring(x[10], 5, 5) , substring(x[10], 5, 6)))

DIS_LEN$variable <- as.numeric(as.character(DIS_LEN$variable))
DIS_LEN$ind_length <- as.numeric(as.character(DIS_LEN$ind_length))

DIS_NO_LEN <- merge(DIS_NO, DIS_LEN, 
                    by=colnames(DIS_LEN)[colnames(DIS_LEN) != "ind_length"], all=T )

DIS_NO_LEN$ind_length[DIS_NO_LEN$ind_length == -1] <- 0

DIS_NO_WEI_LEN <- merge(DIS_NO, merge(DIS_NO_LEN, 
                                      DIS_NO_WEI, all=T), all=T)

DIS_NO_WEI_LEN <- DIS_NO_WEI_LEN[DIS_NO_WEI_LEN$number > 0, ]
DIS_NO_WEI_LEN$NO_AGE_MEASUREMENTS_DISCARDS[DIS_NO_WEI_LEN$NO_AGE_MEASUREMENTS_DISCARDS == -1] <- 0
DIS_NO_WEI_LEN$DISCARDS[DIS_NO_WEI_LEN$DISCARDS == -1] <- 0

kable(data.frame(DIS_Numbers_weights_lengths_from_A_Catch =colnames(DIS_NO_WEI_LEN)))



DIS_NO_WEI_LEN$prod_len <- with(DIS_NO_WEI_LEN, number* ind_length)
DIS_NO_WEI_LEN$prod_wei <- with(DIS_NO_WEI_LEN, number* ind_weight)


colnames(DIS_NO_WEI_LEN)[10:12] <- c("UNWANTED_CATCH","NO_SAMPLES_UC",
                                     "NO_AGE_MEASUREMENTS_UC")


LAN_DIS <- data.frame(DIS_NO_WEI_LEN)

UNWANTED_unique <- group_by(LAN_DIS, COUNTRY, YEAR, QUARTER, VESSEL_LENGTH, 
                            GEAR, MESH_SIZE_RANGE, FISHERY, AREA, SPECIES, UNWANTED_CATCH, 
                            NO_SAMPLES_UC, NO_AGE_MEASUREMENTS_UC)

UNWANTED_unique <- data.frame(summarise(UNWANTED_unique, 
                                        no_records = length(UNWANTED_CATCH) ))

unwanted_catches_sum <- group_by(UNWANTED_unique, COUNTRY, YEAR, QUARTER, VESSEL_LENGTH, 
                                 GEAR, MESH_SIZE_RANGE, FISHERY, AREA, SPECIES)
unwanted_catches_sum <- data.frame(summarise(unwanted_catches_sum, 
                                             total_unwanted_catch = sum(UNWANTED_CATCH) , 
                                             total_unwanted_samples = sum(NO_SAMPLES_UC), 
                                             total_unwanted_age_measurements = sum(NO_AGE_MEASUREMENTS_UC) )) 

unwanted_numbers_len_wei <-  group_by(LAN_DIS, COUNTRY, YEAR, QUARTER, VESSEL_LENGTH, 
                                      GEAR, MESH_SIZE_RANGE, FISHERY, AREA, SPECIES, variable )

unwanted_numbers_len_wei <- data.frame(summarise(unwanted_numbers_len_wei, 
                                                 NO_AGE_UC = sum(number), SUM_PRODS_LEN = sum(prod_len), 
                                                 SUM_PRODS_WEI = sum(prod_wei))) 

unwanted_numbers_len_wei$MEAN_WEIGHT_UC <- with(unwanted_numbers_len_wei,
                                                SUM_PRODS_WEI/NO_AGE_UC)
unwanted_numbers_len_wei$MEAN_LENGTH_UC <- with(unwanted_numbers_len_wei,
                                                SUM_PRODS_LEN/NO_AGE_UC)

C_unwanted_catch_min_max <-  group_by(unwanted_numbers_len_wei, COUNTRY, YEAR, 
                                      QUARTER, VESSEL_LENGTH, 
                                      GEAR,  MESH_SIZE_RANGE, FISHERY,  AREA, 
                                      SPECIES) 

C_UC_min_max <-  data.frame(summarise(C_unwanted_catch_min_max,   
                                      MIN_AGE= min(variable), 
                                      MAX_AGE= max(variable) )) 

C_UC_min_max$MIN_AGE <- round(C_UC_min_max$MIN_AGE, 0)
C_UC_min_max$MAX_AGE <- round(C_UC_min_max$MAX_AGE, 0)

C_UC  <- merge(unwanted_numbers_len_wei, C_UC_min_max, by=c("FISHERY", "COUNTRY" , 
                                                            "YEAR", "QUARTER", "VESSEL_LENGTH", 
                                                            "GEAR", "MESH_SIZE_RANGE",  
                                                            "AREA", "SPECIES") , all=T) 

C_UC_2  <- merge(C_UC, unwanted_catches_sum, by=c( "FISHERY", "COUNTRY" , 
                                                   "YEAR", "QUARTER", "VESSEL_LENGTH", 
                                                   "GEAR", "MESH_SIZE_RANGE",  
                                                   "AREA", "SPECIES") , all.x=T)

#merge with the communication table
colnames(lev5)[2] <- "FISHERY"
C_UC_2 <- merge(C_UC_2, lev5)

C_UC_2$AREA <-  as.numeric(substring(as.character(C_UC_2$AREA) , 4,nchar(as.character(C_UC_2$AREA))))

C_UC_2$DOMAIN_DISCARDS <- paste(C_UC_2$COUNTRY, "_", sep="")
C_UC_2$DOMAIN_DISCARDS  <-  apply(C_UC_2, 1, function(x) ifelse(x[4] == -1, paste(x[22],
                                                                                  "all_", sep=""), paste( x[22], as.numeric(as.character(x[4])),
                                                                                                          "_" , sep="") ) )

C_UC_2$DOMAIN_DISCARDS <- with(C_UC_2, paste(DOMAIN_DISCARDS, "GSA", 
                                             AREA, "_", GEAR, "_", JRC_FDI_codification ,"_", 
                                             MESH_SIZE_RANGE ,  "_NA_NA_" , sep="" )  )   

C_UC_2$DOMAIN_DISCARDS  <-  apply(C_UC_2, 1, function(x) 
  ifelse(x[5] == -1, paste(x[22], "all_", sep=""), 
         paste( x[22], as.numeric(as.character(x[5])),"_" , sep="") ) )

C_UC_2$DOMAIN_DISCARDS <- with(C_UC_2, paste(DOMAIN_DISCARDS, "all_NK" ,
                                             sep="" )  )   

LANDINGS_unique <- group_by(JRC_FDI_Table_E, COUNTRY, SUB_REGION, YEAR, QUARTER, GEAR_TYPE, TARGET_ASSEMBLAGE, DOMAIN_LANDINGS, SPECIES, 
                            TOTWGHTLANDG)
LANDINGS_unique <- data.frame(summarise(LANDINGS_unique, 
                                        no_records = length(TOTWGHTLANDG) ))
#
colnames(LANDINGS_unique)[colnames(LANDINGS_unique)=="DOMAIN_LANDINGS"]<-"DOMAIN_DISCARDS"

C_UC_2[which(as.character(C_UC_2$QUARTER)=="-1"),]$QUARTER="ALL"

#C_UC_3 <- C_UC_2


C_UC_3 <- merge(C_UC_2, LANDINGS_unique)


final_C <- data.frame(COUNTRY  = C_UC_3$COUNTRY   ,
                      YEAR   = C_UC_3$YEAR   ,
                      QUARTER=C_UC_3$QUARTER,
                      SUB_REGION=C_UC_3$AREA,
                      GEAR_TYPE=C_UC_3$GEAR,
                      TARGET_ASSEMBLAGE=C_UC_3$JRC_FDI_codification,
                      DOMAIN_DISCARDS   = C_UC_3$DOMAIN_DISCARDS   ,
                      SPECIES      = C_UC_3$SPECIES   ,
                      TOTWGHTLANDG    = C_UC_3$TOTWGHTLANDG   ,
                      UNWANTED_CATCH    = C_UC_3$ total_unwanted_catch   ,
                      NO_SAMPLES_UC     = C_UC_3$total_unwanted_samples   ,
                      NO_AGE_MEASUREMENTS   = C_UC_3$total_unwanted_age_measurements   ,
                      AGE_MEASUREMENTS_PROP  = 1  ,
                      MIN_AGE   = C_UC_3$MIN_AGE,
                      MAX_AGE   = C_UC_3$MAX_AGE,
                      AGE      = C_UC_3$variable,
                      NO_AGE   = C_UC_3$NO_AGE,
                      MEAN_WEIGHT   = C_UC_3$MEAN_WEIGHT_UC ,
                      MEAN_LENGTH   = C_UC_3$MEAN_LENGTH_UC   
)


final_C$MEAN_WEIGHT <- round(as.numeric(as.character(final_C$MEAN_WEIGHT)) , 3)
final_C$NO_AGE <- round(as.numeric(as.character(final_C$NO_AGE)) , 3)

final_C$MEAN_LENGTH <- round(as.numeric(as.character(final_C$MEAN_LENGTH)) ,1)

final_C <- final_C[with(final_C, order(COUNTRY, YEAR,  SPECIES, DOMAIN_DISCARDS, AGE)), ]
final_C$NO_AGE_MEASUREMENTS[is.na(final_C$NO_AGE_MEASUREMENTS)] <- 0

table_C_props <- final_C 

table_C_props_min_max  <- group_by(final_C, DOMAIN_DISCARDS, SPECIES)
table_C_props_min_max <- data.frame(summarise(table_C_props_min_max, MIN_AGE= min(AGE) , 
                                              MAX_AGE= max(AGE) )) 

table_C_props_ages  <- group_by(table_C_props, SPECIES)
table_C_props_ages <- data.frame(summarise(table_C_props_ages, TOTAL_NO_AGE_MEASUREMENTS= sum(unique(NO_AGE_MEASUREMENTS)))) 

table_C_props <- table_C_props[, !(colnames(table_C_props) %in% c("MIN_AGE", "MAX_AGE") ) ]

final_C <- merge(table_C_props, merge(table_C_props_ages, table_C_props_min_max) )

final_C$AGE_MEASUREMENTS_PROP <- with( final_C,  
                                       NO_AGE_MEASUREMENTS/ TOTAL_NO_AGE_MEASUREMENTS )

final_C$NO_AGE_MEASUREMENTS[final_C$NO_AGE_MEASUREMENTS_UC == 0 ] <- -1
final_C$AGE_MEASUREMENTS_PROP[final_C$TOTAL_NO_AGE_MEASUREMENTS== 0] <- "NK" 

final_C$NO_AGE_MEASUREMENTS <-  final_C$TOTAL_NO_AGE_MEASUREMENTS
final_C$UNWANTED_CATCH <- round(final_C$UNWANTED_CATCH,3)

final_C_2 <- data.frame(COUNTRY = final_C$COUNTRY ,
                        YEAR = final_C$ YEAR,
                        QUARTER=final_C$QUARTER,
                        SUB_REGION=final_C$SUB_REGION,
                        GEAR_TYPE=final_C$GEAR_TYPE,
                        TARGET_ASSEMBLAGE=as.character(final_C$TARGET_ASSEMBLAGE),
                        DOMAIN_DISCARDS = final_C$DOMAIN_DISCARDS ,
                        SPECIES   = final_C$SPECIES,
                        TOTWGHTLANDG  = final_C$TOTWGHTLANDG ,
                        UNWANTED_CATCH   = final_C$UNWANTED_CATCH  ,
                        NO_SAMPLES   = final_C$NO_SAMPLES_UC ,
                        NO_AGE_MEASUREMENTS    = final_C$NO_AGE_MEASUREMENTS,
                        AGE_MEASUREMENTS_PROP = final_C$AGE_MEASUREMENTS_PROP,
                        MIN_AGE     = final_C$MIN_AGE ,
                        MAX_AGE   = final_C$MAX_AGE ,
                        AGE  = final_C$AGE ,
                        NO_AGE   = final_C$NO_AGE ,
                        MEAN_WEIGHT  = final_C$MEAN_WEIGHT,
                        MEAN_LENGTH  = final_C$MEAN_LENGTH )


```


##Output

```{r data_10, eval=T, echo=F}
final_C_2$AGE <- as.numeric(as.character(final_C_2$AGE))
final_C_2 <- final_C_2[with(final_C_2, order(SPECIES, YEAR, DOMAIN_DISCARDS, AGE) ), ]

if(nrow(final_C_2[which(as.character(final_C_2$QUARTER)=="-1"),])>0){
final_C_2[which(as.character(final_C_2$QUARTER)=="-1"),]$QUARTER="ALL"
}

final_C_2$TARGET_ASSEMBLAGE=as.character(final_C_2$TARGET_ASSEMBLAGE)

if(nrow(final_C_2[which(as.character(final_C_2$TARGET_ASSEMBLAGE)=="-1"),])>0){
final_C_2[which(as.character(final_C_2$TARGET_ASSEMBLAGE)=="-1"),]$TARGET_ASSEMBLAGE="NK"
}




colnames(final_C_2)=colnames(template_C)
final_C_2$SUB_REGION=paste("GSA",final_C_2$SUB_REGION,sep="")
write.table(final_C_2, file=paste(getwd(), "/C_DISCARDS_AT_AGE.csv", sep=""), sep=";", row.names=F)

pander(head(final_C_2))
```
